Rebased branch and renamed App_PreChargeSequence to App_Precharge
