+++ boards/BMS/Inc/App/App_BmsWorld.h
@@ -10,7 +10,7 @@
 #include "App_Accumulator.h"
 #include "App_CellMonitors.h"
 #include "App_Airs.h"
-#include "App_PreChargeSequence.h"
+#include "App_PreCharge.h"
 #include "App_SharedErrorTable.h"
 #include "App_SharedClock.h"
