+++ boards/BMS/Src/App/App_PreCharge.c
@@ -1,6 +1,6 @@
 #include <assert.h>
 #include <stdlib.h>
-#include "App_PreChargeSequence.h"
+#include "App_PreCharge.h"
 
 struct PreCharge
 {
